requirements.txt
matplotlib==3.2.2
numpy==1.22.3
opencv_python_headless==4.5.5.64
Pillow==9.4.0
torch==1.11.0
tqdm==4.64.1
